requirements.txt
flask==3.0.0
openai==1.3.0
python-dotenv==1.0.0
flask-dance==7.0.0
oauthlib==3.2.2
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary==2.9.9